Box game scripts done
It is just the xyz factors which needs to be changed accordingly.

notice that the stic needs to be place on the player and to check that
the public game variables have been properly assigned.

t and g, to rotate the stick up and down does not work that well, I
suppose it is because of the camera rotation. That may have to be fixed
at a later point.