Dialogue added in the game! (2 missing!)
All the dialogue textures have been added in a "Dialogue Textures"
folder in Assets.
All dialogue have been added to the right game objects!
But two files are still missing!
Dia#12_line#1 - It was a PSD file and not a PNG file, for some reason
can't I open it.
Dia#19_line#1 - Is broken, inside Unity is it blue...