Fixed subtitle script
So I am not sure how the scripts was suppose to work, but I could not
add the "ShowText" as an input to the "Subtitles" script.
I also found it unnesseary and mearged the two scripts into a new
"Subtitles" script.

So the way the script works is that it is added to an empty object, the
object will then calculate it's Abs distance between itself and he
player, when it gets close enough shall it trigger then text that is
written as an input, for now will it show once for 5 seconds before
disappearing again.