Subtitles added (not all) READ DISCRIPTION!!!
Many things happens in this update so here is the list:
- I removed the skybox! Otherwise couldn't I work so someone have to set
it back when we need it (or maybe just use a nicer default skybox?)
- Subtittles have been added this way:
-- Set an empty object (I called them "Subtitle##")
-- Set what text they should display and after it triggers when it shall
play (default is 0 which is right away)
-- Go into "myStyle" and set fontSize to 24 and Allignment to "Middle
Center"

Comments so far:
I have been using the story doc to find out where the subtitles should
be and I am confused about a few, I added comments in the word dokument
so look there!
The object called "Subtitle03 (Cave, after getting food)" is deactive as
when he returns after getting food shall he say something, so this shall
be activated again when he have got all the food. Also the timers in
that object might need to be fixed at some point. After ALL the
subtittles are done for this object (around 40 seconds of waitting)
shall the choice 3 trigger... not sure how that should be done tho...

The object called "Subtitle04 (Cave, choice 3A)" is also deactivated and
shall be activated if the player decides to stay in the cave (3A). When
this is done shall Carl appear and dialogue 8 shall start.

After "Subtitle 07 (Dying man)" is done (15 sec) then shall the dialogue
13 trigger (again have no clue how this shall be done, maybe a whole new
script with the same kind of Invoke as the subtitle script)

I got to right under dialogue 14 (where they see the boat) and "died
mentally" after that... so I will stop for now, PLEASE read my comments
in the story doc and we can talk about them tomorrow!